refactor(api): renamed proxy/misc package
