+++ api/http/proxy/provider/docker/configs.go
@@ -4,7 +4,7 @@ import (
 	"net/http"
 
 	"github.com/portainer/portainer/api"
-	"github.com/portainer/portainer/api/http/proxy/misc"
+	"github.com/portainer/portainer/api/http/proxy/responseutils"
 )
 
 const (
@@ -22,7 +22,7 @@ func configListOperation(response *http.Response, executor *operationExecutor) e
 
 	// ConfigList response is a JSON array
 	// https://docs.docker.com/engine/api/v1.30/#operation/ConfigList
-	responseArray, err := misc.GetResponseAsJSONArray(response)
+	responseArray, err := responseutils.GetResponseAsJSONArray(response)
 	if err != nil {
 		return err
 	}
@@ -36,7 +36,7 @@ func configListOperation(response *http.Response, executor *operationExecutor) e
 		return err
 	}
 
-	return misc.RewriteResponse(response, responseArray, http.StatusOK)
+	return responseutils.RewriteResponse(response, responseArray, http.StatusOK)
 }
 
 // configInspectOperation extracts the response as a JSON object, verify that the user
@@ -45,7 +45,7 @@ func configListOperation(response *http.Response, executor *operationExecutor) e
 func configInspectOperation(response *http.Response, executor *operationExecutor) error {
 	// ConfigInspect response is a JSON object
 	// https://docs.docker.com/engine/api/v1.30/#operation/ConfigInspect
-	responseObject, err := misc.GetResponseAsJSONOBject(response)
+	responseObject, err := responseutils.GetResponseAsJSONOBject(response)
 	if err != nil {
 		return err
 	}
@@ -57,10 +57,10 @@ func configInspectOperation(response *http.Response, executor *operationExecutor
 	configID := responseObject[configObjectIdentifier].(string)
 	responseObject, access := applyResourceAccessControl(responseObject, configID, executor.operationContext, portainer.ConfigResourceControl)
 	if !access {
-		return misc.RewriteAccessDeniedResponse(response)
+		return responseutils.RewriteAccessDeniedResponse(response)
 	}
 
-	return misc.RewriteResponse(response, responseObject, http.StatusOK)
+	return responseutils.RewriteResponse(response, responseObject, http.StatusOK)
 }
 
 // decorateConfigList loops through all configs and decorates any config with an existing resource control.
