+++ api/http/proxy/provider/docker/networks.go
@@ -4,7 +4,7 @@ import (
 	"net/http"
 
 	"github.com/portainer/portainer/api"
-	"github.com/portainer/portainer/api/http/proxy/misc"
+	"github.com/portainer/portainer/api/http/proxy/responseutils"
 )
 
 const (
@@ -19,7 +19,7 @@ func networkListOperation(response *http.Response, executor *operationExecutor) 
 	var err error
 	// NetworkList response is a JSON array
 	// https://docs.docker.com/engine/api/v1.28/#operation/NetworkList
-	responseArray, err := misc.GetResponseAsJSONArray(response)
+	responseArray, err := responseutils.GetResponseAsJSONArray(response)
 	if err != nil {
 		return err
 	}
@@ -33,7 +33,7 @@ func networkListOperation(response *http.Response, executor *operationExecutor) 
 		return err
 	}
 
-	return misc.RewriteResponse(response, responseArray, http.StatusOK)
+	return responseutils.RewriteResponse(response, responseArray, http.StatusOK)
 }
 
 // networkInspectOperation extracts the response as a JSON object, verify that the user
@@ -42,7 +42,7 @@ func networkListOperation(response *http.Response, executor *operationExecutor) 
 func networkInspectOperation(response *http.Response, executor *operationExecutor) error {
 	// NetworkInspect response is a JSON object
 	// https://docs.docker.com/engine/api/v1.28/#operation/NetworkInspect
-	responseObject, err := misc.GetResponseAsJSONOBject(response)
+	responseObject, err := responseutils.GetResponseAsJSONOBject(response)
 	if err != nil {
 		return err
 	}
@@ -54,30 +54,30 @@ func networkInspectOperation(response *http.Response, executor *operationExecuto
 	networkID := responseObject[networkIdentifier].(string)
 	responseObject, access := applyResourceAccessControl(responseObject, networkID, executor.operationContext, portainer.NetworkResourceControl)
 	if access {
-		return misc.RewriteResponse(response, responseObject, http.StatusOK)
+		return responseutils.RewriteResponse(response, responseObject, http.StatusOK)
 	}
 
 	networkLabels := extractNetworkLabelsFromNetworkInspectObject(responseObject)
 	responseObject, access = applyResourceAccessControlFromLabel(networkLabels, responseObject, networkLabelForStackIdentifier, executor.operationContext, portainer.StackResourceControl)
 	if access {
-		return misc.RewriteResponse(response, responseObject, http.StatusOK)
+		return responseutils.RewriteResponse(response, responseObject, http.StatusOK)
 	}
 
-	return misc.RewriteAccessDeniedResponse(response)
+	return responseutils.RewriteAccessDeniedResponse(response)
 }
 
 // extractNetworkLabelsFromNetworkInspectObject retrieve the Labels of the network if present.
 // Container schema reference: https://docs.docker.com/engine/api/v1.28/#operation/NetworkInspect
 func extractNetworkLabelsFromNetworkInspectObject(responseObject map[string]interface{}) map[string]interface{} {
 	// Labels are stored under Labels
-	return misc.ExtractJSONField(responseObject, "Labels")
+	return responseutils.GetJSONObject(responseObject, "Labels")
 }
 
 // extractNetworkLabelsFromNetworkListObject retrieve the Labels of the network if present.
 // Network schema reference: https://docs.docker.com/engine/api/v1.28/#operation/NetworkList
 func extractNetworkLabelsFromNetworkListObject(responseObject map[string]interface{}) map[string]interface{} {
 	// Labels are stored under Labels
-	return misc.ExtractJSONField(responseObject, "Labels")
+	return responseutils.GetJSONObject(responseObject, "Labels")
 }
 
 // decorateNetworkList loops through all networks and decorates any network with an existing resource control.
