+++ api/http/proxy/provider/docker/volumes.go
@@ -4,7 +4,7 @@ import (
 	"net/http"
 
 	"github.com/portainer/portainer/api"
-	"github.com/portainer/portainer/api/http/proxy/misc"
+	"github.com/portainer/portainer/api/http/proxy/responseutils"
 )
 
 const (
@@ -19,7 +19,7 @@ func volumeListOperation(response *http.Response, executor *operationExecutor) e
 	var err error
 	// VolumeList response is a JSON object
 	// https://docs.docker.com/engine/api/v1.28/#operation/VolumeList
-	responseObject, err := misc.GetResponseAsJSONOBject(response)
+	responseObject, err := responseutils.GetResponseAsJSONOBject(response)
 	if err != nil {
 		return err
 	}
@@ -42,7 +42,7 @@ func volumeListOperation(response *http.Response, executor *operationExecutor) e
 		responseObject["Volumes"] = volumeData
 	}
 
-	return misc.RewriteResponse(response, responseObject, http.StatusOK)
+	return responseutils.RewriteResponse(response, responseObject, http.StatusOK)
 }
 
 // volumeInspectOperation extracts the response as a JSON object, verify that the user
@@ -51,7 +51,7 @@ func volumeListOperation(response *http.Response, executor *operationExecutor) e
 func volumeInspectOperation(response *http.Response, executor *operationExecutor) error {
 	// VolumeInspect response is a JSON object
 	// https://docs.docker.com/engine/api/v1.28/#operation/VolumeInspect
-	responseObject, err := misc.GetResponseAsJSONOBject(response)
+	responseObject, err := responseutils.GetResponseAsJSONOBject(response)
 	if err != nil {
 		return err
 	}
@@ -63,30 +63,30 @@ func volumeInspectOperation(response *http.Response, executor *operationExecutor
 	volumeID := responseObject[volumeIdentifier].(string)
 	responseObject, access := applyResourceAccessControl(responseObject, volumeID, executor.operationContext, portainer.VolumeResourceControl)
 	if access {
-		return misc.RewriteResponse(response, responseObject, http.StatusOK)
+		return responseutils.RewriteResponse(response, responseObject, http.StatusOK)
 	}
 
 	volumeLabels := extractVolumeLabelsFromVolumeInspectObject(responseObject)
 	responseObject, access = applyResourceAccessControlFromLabel(volumeLabels, responseObject, volumeLabelForStackIdentifier, executor.operationContext, portainer.StackResourceControl)
 	if access {
-		return misc.RewriteResponse(response, responseObject, http.StatusOK)
+		return responseutils.RewriteResponse(response, responseObject, http.StatusOK)
 	}
 
-	return misc.RewriteAccessDeniedResponse(response)
+	return responseutils.RewriteAccessDeniedResponse(response)
 }
 
 // extractVolumeLabelsFromVolumeInspectObject retrieve the Labels of the volume if present.
 // Volume schema reference: https://docs.docker.com/engine/api/v1.28/#operation/VolumeInspect
 func extractVolumeLabelsFromVolumeInspectObject(responseObject map[string]interface{}) map[string]interface{} {
 	// Labels are stored under Labels
-	return misc.ExtractJSONField(responseObject, "Labels")
+	return responseutils.GetJSONObject(responseObject, "Labels")
 }
 
 // extractVolumeLabelsFromVolumeListObject retrieve the Labels of the volume if present.
 // Volume schema reference: https://docs.docker.com/engine/api/v1.28/#operation/VolumeList
 func extractVolumeLabelsFromVolumeListObject(responseObject map[string]interface{}) map[string]interface{} {
 	// Labels are stored under Labels
-	return misc.ExtractJSONField(responseObject, "Labels")
+	return responseutils.GetJSONObject(responseObject, "Labels")
 }
 
 // decorateVolumeList loops through all volumes and decorates any volume with an existing resource control.
